fix(test): cr delete api doesn't return nil if the cr is not found
The new CR delete api doesn't return nil if the cr is not found, so
we cannot rely on the nil value to check if the cr exists.

Longhorn 7944

+++ controller/node_controller_test.go
@@ -1902,7 +1902,7 @@ func (s *NodeControllerSuite) TestSyncInstanceManagers(c *C) {
 			if err != nil && !apierrors.IsNotFound(err) {
 				c.Assert(err, IsNil)
 			}
-			if im != nil {
+			if im != nil && err == nil {
 				err := s.lhClient.LonghornV1beta2().InstanceManagers(TestNamespace).Delete(context.TODO(), instanceManager.Name, metav1.DeleteOptions{})
 				c.Assert(err, IsNil)
 			}
